backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-decouple
python-jose
passlib[bcrypt]
cryptography
